add bigint support

+++ tests/TabSample.sql
@@ -1,5 +1,5 @@
 CREATE TABLE tab_bar (
-  alpha int  ,
+  alpha bigint  ,
   beta varchar(255) DEFAULT NULL,
   gamma int,
   delta varchar(255) DEFAULT NULL
